fetcher to global stage
