+++ src/pages/_app.tsx
@@ -1,12 +1,23 @@
-import 'tailwindcss/tailwind.css'
-// import '@/styles/globals.css'
+// import 'tailwindcss/tailwind.css'
+import '@/styles/globals.css'
 import React from 'react'
+import { SWRConfig } from 'swr'
 
 import { useBgColor } from '@/hooks/useBgColor'
 import { useCounter } from '@/hooks/useCounter'
 import { useInputArray } from '@/hooks/useInputArray'
 
 import type { AppProps } from 'next/app'
+
+const fetcher = async (url: any) => {
+  const response = await fetch(url)
+
+  if (!response.ok) {
+    throw new Error('エラー')
+  }
+  const json = await response.json()
+  return json
+}
 
 const App = ({ Component, pageProps }: AppProps) => {
   const counter = useCounter()
@@ -15,7 +26,11 @@ const App = ({ Component, pageProps }: AppProps) => {
 
   useBgColor()
 
-  return <Component {...pageProps} {...counter} {...inputArray} />
+  return (
+    <SWRConfig value={{ fetcher }}>
+      <Component {...pageProps} {...counter} {...inputArray} />
+    </SWRConfig>
+  )
 }
 
 export default App
