+++ tailwind.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   mode: 'jit',
-  purge: ['./src/**/*.{js,ts,jsx,tsx}'],
+  // purge: ['./src/**/*.{js,ts,jsx,tsx}'],
   theme: {
     extend: {}
   },
